Reorganizando el proyecto.
- Se creo el componente balance.
- Se reorganizo el arbol de archivos.
- Se establecio la ruta /board como ruta principal temporalmente.
- Se agregaron medidas relativas a los componentes principales.
