+++ src/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import FinanceBoard from './components/pages/FinanceBoard'
+import FinanceBoard from './components/pages/financeBoard'
 import Login from './components/pages/login'
 import { BrowserRouter as Router, Route } from 'react-router-dom'
 
@@ -8,8 +8,9 @@ const App = () => {
 
     return (
         <Router >
-            <Route exact path = "/" component = { Login }/>
-            <Route path = "/board" component = { FinanceBoard } />
+            <Route exact path="/" component={ FinanceBoard }/>
+            <Route path="/board" component={ FinanceBoard } />
+            <Route path="/pruebas" component={ Login } />
         </Router>
     )
 }
