+++ src/components/organisms/balance/balance.css
@@ -1,0 +1,23 @@
+.balance {
+    color: #fff;
+    height: 20%;
+    width: 100%;
+    display: flex;
+    justify-content: center; 
+    margin-bottom: 50px;
+}
+
+.content-item-balance {
+    padding: 40px;
+    text-align: center;
+}
+
+.content-icome {
+    background-color: rgb(93,192,97, 0.5);
+    border-radius: 0px 0px 0px 10px;
+}
+
+.content-expenses {
+    background-color: rgb(229,61,47, 0.5);
+    border-radius: 0px 0px 10px 0px;
+}
