+++ src/components/organisms/transaction/transaction.css
@@ -4,7 +4,7 @@
     display: flex;
     justify-content: space-around;
     transition: 450ms;
-    width: 400px;
+    width: 95%;
     margin-left: 15px;
 }
