--- src/components/pages/financeBoard.css
@@ -1,3 +1,0 @@
-.bodyPages   {
-    display: flex;
-}
