+++ src/components/pages/financeBoard/financeBoard.css
@@ -1,0 +1,15 @@
+.bodyPages   {
+    display: flex;
+    justify-content: space-around;
+}
+
+
+@media (max-width: 600px) {
+    .bodyPages {
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+    }
+  }
+
