+++ src/index.css
@@ -2,7 +2,6 @@ body {
   background-color: rgb(10,10,10);
   margin: 0px;
   font-family: 'Josefin Sans', sans-serif;
-  /* overflow-y: hidden; */
 }
 
 span {
