+++ src/index.js
@@ -6,9 +6,7 @@ import App from './App';
 
 ReactDOM.render(
   <React.StrictMode>
-   
     <App /> 
-
   </React.StrictMode>,
   
   document.getElementById('root')
